python-toolkit/requirements.txt
python-nmap>=0.7.1
pexpect>=4.8.0
ipaddress>=1.0.23
